+++ client/src/LanguageClient/LocalLanguageClientProvider.ts
@@ -7,7 +7,9 @@ import {
     LanguageClient,
     LanguageClientOptions,
     TransportKind,
-    ServerOptions
+    ServerOptions,
+    ErrorAction,
+    CloseAction
 } from 'vscode-languageclient'
 
 import * as path from 'path'
@@ -46,10 +48,16 @@ export class LocalLanguageClientProvider {
                 configurationSection: "openhab",
                 fileEvents: workspace.createFileSystemWatcher("**/.clientrc"),
             },
+            // Disable the default error handler
+            errorHandler: {
+                error: () => ErrorAction.Continue,
+                closed: () => CloseAction.DoNotRestart
+            }
         }
 
         // Create the language client and start the client.
         const lc = new LanguageClient("openhabLanguageServer", "Openhab Language Server", serverOptions, clientOptions)
         return lc.start()
     }
+
 }